remote_access: Fix ping seqno type.
esp_ping_get_profile complained about
"E (70983) ping_sock: esp_ping_get_profile(439):
 unmatched data size for profile 0"

+++ software/src/modules/remote_access/remote_access.cpp
@@ -1945,7 +1945,7 @@ void RemoteAccess::close_all_remote_connections() {
 
 static void on_ping_success(esp_ping_handle_t handle, void *args) {
     uint8_t ttl;
-    uint8_t seqno;
+    uint16_t seqno;
     uint32_t elapsed_time, recv_len;
     ip_addr_t target_addr;
 
@@ -1972,7 +1972,7 @@ static void on_ping_success(esp_ping_handle_t handle, void *args) {
 }
 
 static void on_ping_timeout(esp_ping_handle_t handle, void *args) {
-    uint8_t seqno;
+    uint16_t seqno;
     ip_addr_t target_addr;
 
     esp_ping_get_profile(handle, ESP_PING_PROF_SEQNO, &seqno, sizeof(seqno));
